# api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Dim active API", "version": "1.0.0" },
  "servers": [{ "url": "/api/v1" }],
  "paths": {
...
    "/media/{id}": { "parameters": [{ "$ref": "#/components/parameters/Id" }], "get": { "operationId": "getMedia", "security": [{ "token": [] }], "responses": { "200": { "description": "Media", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Media" } } } }, "404": { "$ref": "#/components/responses/Error" } } } },
...
    "/media/{id}/progress": { "parameters": [{ "$ref": "#/components/parameters/Id" }], "post": { "operationId": "saveProgress", "security": [{ "token": [] }], "parameters": [{ "name": "offset", "in": "query", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Saved" } } } },
    "/media/tmdb_search": { "get": { "operationId": "searchExternalMedia", "security": [{ "token": [] }], "parameters": [{ "name": "query", "in": "query", "required": true, "schema": { "type": "string" } }, { "name": "media_type", "in": "query", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Results", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ExternalMedia" } } } } } } } },
...
  },
  "components": {
    "securitySchemes": { "token": { "type": "apiKey", "in": "header", "name": "Authorization" }, "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "dim_session" } },
    "parameters": { "Id": { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }, "Gid": { "name": "gid", "in": "path", "required": true, "schema": { "type": "string" } } },
...
    "responses": { "Login": { "description": "Token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } } }, "Register": { "description": "Created account and token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterResponse" } } } }, "Error": { "description": "Structured error", "headers": { "x-request-id": { "schema": { "type": "string" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorEnvelope" } } } } },
    "schemas": {
      "ApiErrorEnvelope": { "type": "object", "required": ["error", "request_id"], "properties": { "error": { "$ref": "#/components/schemas/ApiError" }, "request_id": { "type": "string" } } },
      "ApiError": { "type": "object", "required": ["code", "message"], "properties": { "code": { "type": "string" }, "message": { "type": "string" }, "details": { "type": "object", "additionalProperties": true } } },
...
      "LoginResponse": { "type": "object", "required": ["token"], "properties": { "token": { "type": "string" } } },
      "RegisterResponse": { "type": "object", "required": ["token", "username"], "properties": { "token": { "type": "string" }, "username": { "type": "string" } } },
...
      "Chapters": { "type": "object", "required": ["credits"], "properties": { "credits": { "type": "integer" } } },
      "Media": { "type": "object", "required": ["id", "name", "library_id", "media_type", "genres", "duration", "progress"], "properties": { "id": { "type": "integer" }, "name": { "type": "string" }, "library_id": { "type": "integer" }, "media_type": { "type": "string" }, "genres": { "type": "array", "items": { "type": "string" } }, "duration": { "type": "integer" }, "progress": { "type": "integer" }, "description": { "type": ["string", "null"] }, "added": { "type": ["string", "null"] }, "backdrop_path": { "type": ["string", "null"] }, "poster_path": { "type": ["string", "null"] }, "chapters": { "$ref": "#/components/schemas/Chapters" }, "episode": { "type": "integer" }, "next_episode_id": { "type": "integer" }, "play_btn_id": { "type": "integer" }, "prev_episode_id": { "type": "integer" }, "rating": { "type": "number" }, "season": { "type": "integer" }, "year": { "type": "integer" }, "tags": { "type": "object", "additionalProperties": true } }, "additionalProperties": true },
...
      "ExternalMedia": { "type": "object", "additionalProperties": true },
...
    }
  }
}